change MyProjectSelect component's name to "ProjectList" (more consistent)
